Move to own file that can be build tagged
