+++ set.go
@@ -35,11 +35,6 @@ SOFTWARE.
 // that can enforce mutual exclusion through other means.
 package mapset
 
-import (
-	"cmp"
-	"slices"
-)
-
 // Set is the primary interface provided by the mapset package.  It
 // represents an unordered set of data and a large number of
 // operations that can be applied to that set.
@@ -246,13 +241,5 @@ func NewThreadUnsafeSetFromMapKeys[T comparable, V any](val map[T]V) Set[T] {
 		s.Add(k)
 	}
 
-	return s
-}
-
-// Sorted returns a sorted slice of a set of any ordered type in ascending order.
-// When sorting floating-point numbers, NaNs are ordered before other values.
-func Sorted[E cmp.Ordered](set Set[E]) []E {
-	s := set.ToSlice()
-	slices.Sort(s)
 	return s
 }
