+++ set_test.go
@@ -1298,34 +1298,6 @@ func Test_NewThreadUnsafeSetFromMapKey_Strings(t *testing.T) {
 	}
 }
 
-func Test_Sorted(t *testing.T) {
-	test := func(t *testing.T, ctor func(vals ...string) Set[string]) {
-		set := ctor("apple", "banana", "pear")
-		sorted := Sorted(set)
-
-		if len(sorted) != set.Cardinality() {
-			t.Errorf("Length of slice is not the same as the set. Expected: %d. Actual: %d", set.Cardinality(), len(sorted))
-		}
-
-		if sorted[0] != "apple" {
-			t.Errorf("Element 0 was not equal to apple: %s", sorted[0])
-		}
-		if sorted[1] != "banana" {
-			t.Errorf("Element 0 was not equal to banana: %s", sorted[1])
-		}
-		if sorted[2] != "pear" {
-			t.Errorf("Element 0 was not equal to pear: %s", sorted[2])
-		}
-	}
-
-	t.Run("Safe", func(t *testing.T) {
-		test(t, NewSet[string])
-	})
-	t.Run("Unsafe", func(t *testing.T) {
-		test(t, NewThreadUnsafeSet[string])
-	})
-}
-
 func Test_Example(t *testing.T) {
 	/*
 	   requiredClasses := NewSet()
